requirements.txt
Flask==3.0.3
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
openpyxl==3.1.5
xlsxwriter==3.2.0
pytest==8.3.3
python_docx==1.1.2
setuptools==69.5.1
tqdm==4.66.4
Werkzeug==3.0.4
scikit-learn==1.3.2